+++ src/babel-plugin-dynamic-import-amd.ts
@@ -66,8 +66,12 @@ export const dynamicImportAmd = {
       // Transform the dynamic import callsites
       for (const importPath of dynamicImports) {
         const specifier = importPath.node.arguments[0];
+        // Call as `require.default` because the AMD transformer that we assume
+        // is running next will rewrite `require` from a function to a module
+        // object with the function at `default`.
         importPath.replaceWith(ast`(
-          new Promise((res, rej) => ${requireId}([${specifier}], res, rej))
+          new Promise((res, rej) => ${requireId}.default([${
+            specifier}], res, rej))
         )`);
       }
     },